Keyed use softIk attribute bug
Fix bug where when the use soft ik was keyed it would break or freeze
the stretch behavior.

This issue was in animCurve where when adding a key we would force the
value to be converted to a float regardless it’s type. In the case of a
boolean this would get converted in Nan.